requirements.txt
attrs==22.1.0
Automat==22.10.0
cassandra-driver==3.25.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
cryptography==38.0.4
cssselect==1.2.0
filelock==3.8.0
geomet==0.2.1.post1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.1
packaging==21.3
parsel==1.7.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.7.1
service-identity==21.1.0
six==1.16.0
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
w3lib==2.1.0
zope.interface==5.5.2
